+++ src/lib/remark-asset-images.ts
@@ -3,26 +3,11 @@ import { existsSync } from "node:fs";
 import path from "node:path";
 import type { Root } from "mdast";
 import type { VFile } from "vfile";
+import { tinaCloudMediaPath } from "./tinaImageUrl";
 
 // Tina's media root (see tina/config.ts `media.tina.mediaRoot`). Cloud asset
-// URLs encode the path *relative to this folder* after the `/__file/` marker.
+// URLs encode the path *relative to this folder*.
 const MEDIA_ROOT = "src/assets";
-
-// Cloud builds rewrite body-image URLs to an absolute Tina Cloud CDN URL. Two
-// forms occur: staging `.../<id>/__staging/<branch>/__file/<media-root-path>`
-// and published `.../<id>/<media-root-path>`.
-const TINA_CLOUD_FILE = /^https?:\/\/assets\.tina\.io\/.*\/__file\/(.+)$/;
-const TINA_CLOUD_ASSET = /^https?:\/\/assets\.tina\.io\/[^/]+\/(.+)$/i;
-
-// Reduce either Tina Cloud asset URL form to its media-root-relative path.
-function tinaMediaPath(url: string): string | undefined {
-  const file = url.match(TINA_CLOUD_FILE);
-  if (file) {
-    return file[1];
-  }
-  const asset = url.match(TINA_CLOUD_ASSET);
-  return asset ? asset[1] : undefined;
-}
 
 export function remarkAssetImages() {
   return (tree: Root, file: VFile) => {
@@ -48,7 +33,7 @@ export function remarkAssetImages() {
       // Cloud build: reduce the Tina Cloud URL back to its media-root-relative
       // path and resolve it against `src/assets`, otherwise the page would
       // hot-link to Tina Cloud (see scripts/test/no-tina-cloud-urls.ts).
-      let mediaRelative = tinaMediaPath(node.url);
+      let mediaRelative = tinaCloudMediaPath(node.url);
       if (mediaRelative) {
         try {
           mediaRelative = decodeURIComponent(mediaRelative);
